Update Hive overrides
