+++ scripts/rundiffspark2.sh
@@ -50,6 +50,9 @@ if [[ $(diff spark2diffs/GpuHiveOverrides.diff $tmp_dir/GpuHiveOverrides.newdiff
   echo "check diff for  ../spark2-sql-plugin/src/main/scala/org/apache/spark/sql/hive/rapids/GpuHiveOverrides.scala"
 fi
 
+diff ../sql-plugin/src/main/scala/org/apache/spark/sql/hive/rapids/HiveProviderImpl.scala ../spark2-sql-plugin/src/main/scala/org/apache/spark/sql/hive/rapids/HiveProviderImpl.scala > $tmp_dir/HiveProviderImpl.newdiff || true
+diff -c spark2diffs/HiveProviderImpl.diff $tmp_dir/HiveProviderImpl.newdiff
+
 sed -n  '/class GpuBroadcastNestedLoopJoinMeta/,/override def convertToGpu/{/override def convertToGpu/!p}'  ../spark2-sql-plugin/src/main/scala/org/apache/spark/sql/rapids/execution/GpuBroadcastNestedLoopJoinMeta.scala > $tmp_dir/GpuBroadcastNestedLoopJoinMeta_new.out
 sed -n  '/class GpuBroadcastNestedLoopJoinMeta/,/override def convertToGpu/{/override def convertToGpu/!p}'  ../sql-plugin/src/main/scala/org/apache/spark/sql/rapids/execution/GpuBroadcastNestedLoopJoinExec.scala > $tmp_dir/GpuBroadcastNestedLoopJoinMeta_old.out
 diff $tmp_dir/GpuBroadcastNestedLoopJoinMeta_new.out $tmp_dir/GpuBroadcastNestedLoopJoinMeta_old.out > $tmp_dir/GpuBroadcastNestedLoopJoinMeta.newdiff || true
